MinMaxAnnotationHandler: use non-deprecated BigDecimalRangeRandomizer constructor

+++ random-beans-validation/src/main/java/io/github/benas/randombeans/validation/MinMaxAnnotationHandler.java
@@ -100,8 +100,8 @@ class MinMaxAnnotationHandler implements BeanValidationAnnotationHandler {
             }
             if (fieldType.equals(BigDecimal.class)) {
                 return new BigDecimalRangeRandomizer(
-                        minValue == null ? null : minValue,
-                        maxValue == null ? null : maxValue,
+                        minValue == null ? null : minValue.doubleValue(),
+                        maxValue == null ? null : maxValue.doubleValue(),
                         random.nextLong()
                 );
             }
